Bug/fix captcha (#43)
* Fix: captcha

* fix: verify captcha

* fix: captcha

* update: fix cors, accesskey

* Update: fix invite friend

+++ core_service/aws_lambda/project/code/login_refresh_token.py
@@ -72,7 +72,7 @@ def lambda_handler(event, context):
         "identity_id": identity["IdentityId"],
         "credential_token_expires_in": (identity["Credentials"]["Expiration"].timestamp())*1000, # expire time in seconds
         "id_token": id_token,
-        "token_expires_in": (datetime.now().timestamp() + ACCESS_TOKEN_EXPIRATION)*1000,
+        "token_expires_in":float(int((datetime.now().timestamp() + ACCESS_TOKEN_EXPIRATION)*1000)),
         "username": username
     }
